Added more tests

+++ module/Application/src/Mapper/BaseMapper.php
@@ -5,6 +5,7 @@ namespace Application\Mapper;
 use Doctrine\DBAL\Connection;
 use Doctrine\ORM\{
     EntityManager,
+    EntityNotFoundException,
     EntityRepository,
     OptimisticLockException,
     Exception\ORMException,
@@ -81,13 +82,19 @@ abstract class BaseMapper
      * Remove an entity by its ID using find
      *
      * @param mixed $id
+     * @throws EntityNotFoundException
+     * @throws OptimisticLockException
      * @throws ORMException
      */
     public function removeById(mixed $id): void
     {
         $entity = $this->find($id);
-        $this->em->remove($entity);
-        $this->em->flush();
+        if (!is_null($entity)) {
+            $this->em->remove($entity);
+            $this->em->flush();
+        } else {
+            throw new EntityNotFoundException('No entity with the given ID could be found.');
+        }
     }
 
     /**
